Highlight "lexical differences ignored" for grammar coverage more obviously like earlier drafts (#34)
For space constraints the figure with "lexical differences ignored"
(because our RASP model uses an embedding which transforms words to
part-of-speech / verb tokens) for grammar coverage written all over it
moved to the appendix and this phrase also moved out of the first
results section to Appendix 11 and Appendix 7, so restore that.

Fix a reference to the end of Appendix 7 (Computing Grammar Coverage)
which should be 11 (Grammar Coverage analysis to develop and justify
Restricted Access Sequence Processing model design).


https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/9ec12a16528d92d6770edcae6ffbdc0da55c0e1f/rasp-for-recogs_pos-wbruns-2024-draft.pdf