In project limitations, where ReCOGS vs COGS is mentioned, add a link to the new RASP-for-COGS project undergoing evaluation now where preliminary data is consistent with conclusions in this paper (#44)
Add a link to the new RASP-for-COGS project undergoing evaluation now
where preliminary data is consistent with conclusions in this paper.

PDF preview link:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/670ff335fb98e902f0fc246b9b08698a07b25991/rasp-for-recogs_pos-wbruns-2024-draft.pdf

Also update the existing link to RASP-for-COGS in the README.md to note
it is no longer in development and is undergoing evaluation now.

**Also some style cleanup already done in the deidentified (for review)
copy of the paper that had not been pulled back here (fix linebreaks
before footnotes). This is most of the apparent diff, it looks like a
lot of changes occurred when it is just linebreaking long lines with
footnotes.**

Also add a couple of sentences mentioning the multidimensional vector
interpretation of the 1 dimensional part of speech and verb type tokens
to the appendix.