more cleanup and cite additional related papers (resolve https://github.com/willy-b/RASP-for-ReCOGS/issues/43 )
Fix typographical errors, remove outdated warnings (like a note in the RASP model design appendix noting addition of complement phrase support would change it when in other places it is already noted that the explanation does not cover complement phrase handling and links to the code for that - leftover from first draft before complement phrase handling was supported), add another link to the RASP-for-COGS project in the introduction just to give context, cite a few more relevant papers, remove the word "token" from a figure where it is confusing (see https://github.com/willy-b/RASP-for-ReCOGS/pull/44#discussion_r2014955591 for explanation ). Specifically remove the word "token" referring to an embedded word-level vocabulary token as a "part-of-speech / verb-type token". The embedding is a scalar part-of-speech/verb-type embedding in RASP so is a sequence of scalars of the same form as the input word-level token sequence (and is in fact equivalent to just starting with a token sequence of parts of speech / verb types instead of words), so both descriptions are accurate but I worry putting that into the figure may be confusing.

Since our paper argues by construction that Transformers can do the difficult structural generalizations in the ReCOGS task and the problem is getting them to learn the right rules out of all of those that they can learn to explain the training data, add a citation of Ahuja et al 2025 in conclusion as an example of Transformers learning explanations for the training data with different generalization properties based on different training objectives over same data.