More cleanup and cite additional related papers (resolve https://github.com/willy-b/RASP-for-ReCOGS/issues/43 ) (#45)
PDF preview link:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/0ed5765bea7ad557be42bc2c442571b5f136a8f6/rasp-for-recogs_pos-wbruns-2024-draft.pdf

- Fix typographical errors
- Remove outdated warnings (like a note in the RASP model design
appendix noting addition of complement phrase support would change it
when in other places it is already noted that the explanation does not
cover complement phrase handling and links to the code for that -
leftover from first draft before complement phrase handling was
supported)
- Add another link to the RASP-for-COGS project in the introduction just
to give context
- Cite a few more relevant papers
- Remove the word "token" from a figure where it is confusing (see
https://github.com/willy-b/RASP-for-ReCOGS/pull/44#discussion_r2014955591
for explanation ). Specifically: Remove the word "token" referring to an
embedded word-level vocabulary token as a "part-of-speech / verb-type
token". The embedding is a scalar part-of-speech/verb-type embedding in
RASP so is a sequence of scalars of the same form as the input
word-level token sequence (and is in fact equivalent to just starting
with a token sequence of parts of speech / verb types instead of words),
so both descriptions are accurate but I worry putting that into the
figure may be confusing.

- Since our paper argues by construction that Transformers can do the
difficult structural generalizations in the ReCOGS task and the problem
is getting them to learn the right rules out of all of those that they
can learn to explain the training data, add a citation of Ahuja et al
2025 in conclusion as an example of Transformers learning explanations
for the training data with different generalization properties based on
different training objectives over same data.

- Replace 19 sentences illustrating flat patterns (+2 illustrating flat
rules to handle recursive aspects of grammar) matched with their
part-of-speech/verb-type representation as more informative and less
distracting in results

- Add links to existing relevant figures in the very first sentence of
that results section

- Clarify very old model detail appendix text that could potentially be
misinterpreted as us actually implementing the COGS PCFG
(tree-structured, hierarchical, recursive) in the RASP model rather than
implementing 19 flat pattern matches + 2 flat rules to handle recursive
aspects of the grammar , where those flat rules were generated to cover
(using grammar coverage analysis) the COGS PCFG (as described by Klinger
et al in their utilities, not using Klinger's model but description of
COGS input PCFG, which is the same as ReCOGS input PCFG as the input is
the same, only LFs different)

- Remove from supplemental materials a redundant (Re)COGS input grammar
description (Kim/Linzen grammar described by Klinger in training data
compatible way) used for grammar coverage analysis and in planning
original model design as our RASP model is NOT a PCFG and someone can
just read their utilities/paper and we already have a nonterminals only
equivalent description in the grammar coverage appendix

- In Figure 3 caption add "((d) rejects H0 )" to clarify exactly why the
flat/non-tree structured/non-hierarchical hypothesis is supported in the
figure as it shows a tree-structured representation in the figure which
is then rejected by the observation at the end (we are rejecting the
tree-structured hypothesis with observation (d), leaving our alternative
hypothesis of a flat model ). nobody complained about this but I felt it
was unclear if one did not read the text and just looked at the figures
alone.

- Expand existing note that part-of-speech as earliest representation in
our RASP model was inspired by Tenney et al 2019 finding such
information extracted by the BERT model naturally after learning using
its masked language model objective to note that though COGS/ReCOGS may
be treated as seq2seq problems one could also train the model to predict
both the input and the output (language modeling objective vs seq2seq
were compared by e.g. (Ahuja et al., 2024) for a variety of
generalization problems and they found models trained on the same
problem with a language modeling objective generalized better)