Prepare deidentified version of latest in main branch including PR#45 (#47)
Also replace part of one sentence in the conclusion to focus on an
important finding of the paper already highlighted in abstract and
introduction that was not mentioned again.

Also emphasize ReCOGS_pos as dataset studied in Data section, adding
mention that COGS have the same input sentences/grammar/vocabulary
(apart from ReCOGS augmentations which only occurs in training, not
test/gen). Note that for predicting the original COGS LF from the input
sentences, a RASP-for-COGS solution was split out into a separate
project (RASP-for-COGS) which is under evaluation now.

See https://github.com/willy-b/RASP-for-ReCOGS/pull/45 for a description
of all the changes.

Preview link:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/91e11dccbfb43f2e148fc8bf24cb7fc16777dcbf/rasp-for-recogs_pos-2024-draft.pdf