Improve 2 simplified RASP snippets in appendix: show both verb relationship left AND right index computation, use consistent convention with final 2024-11 RASP code in GitHub, not 2024-10 convention, remove oversimplified position / offset calculation (let reader visit repo for it)
The appendix has some RASP code examples (NOT the actual code used which is referenced via link to the original evaluated versions in github, but simplified snippets).
For the ignoring attractor nouns example showing verb left,right indices, the left-index computation was not shown, and the right-index is not written using same conventions as 2024-11 final RASP code (was in line with 2024-10 style), so adding the left index computation to the example code (simplified vs actual program) and update to conventions of 2024-11 code. I am also removing a line which computes a simplified version of the position/offset because it is misleading compared to what was actually used (which is more complicated) and does not help with understanding. They can read the actual code for that.

Added the left index computation to the main RASP example of verb relationship indices in RASP, but removed the left-index reference from a shorter snippet in a separate section only focused on how the relationship right index avoids attraction errors to intervening prepositional phrase nouns.