Cleanup appendix RASP examples (not actual evaluated code) - Propagate https://github.com/willy-b/RASP-for-ReCOGS/pull/56 appendix changes to deidentified branch (for blind review by conference/journal - not to keep anything secret; OK to be hosted here alongside everything with my name) (#57)
PDF Preview link at:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/2dd621a3aadbbf445e189a7855637b0473f5871e/rasp-for-recogs_pos-2024-draft.pdf

See https://github.com/willy-b/RASP-for-ReCOGS/pull/56 , same thing but
in deidentified (blind review) branch here.

The appendix has some RASP code examples (NOT the actual code used which
is referenced via link to the original evaluated versions in GitHub
(learning-rasp repo, code unchanged since 2024-11), but simplified
snippets).

For the ignoring attractor nouns example showing verb left,right
indices, the left-index computation was not shown, and the right-index
is not written using same conventions as 2024-11 final RASP code (was in
line with 2024-10 style, not enforcing v_inf LF output in order, out of
order allowed by Semantic Exact Match), so adding the left index
computation to the example code (simplified vs actual program) and
update to conventions of 2024-11 code. I am also removing a line which
computes a simplified version of the position/offset because it is
misleading compared to what was actually used (which is more
complicated) and does not help with understanding. They can read the
actual code for that.

Added the left index computation to the main RASP example of verb
relationship indices in RASP, but removed the left-index reference from
a shorter snippet in a separate section only focused on how the
relationship right index avoids attraction errors to intervening
prepositional phrase nouns.

Also updated the appendix to mention how Decoder variables for all
phases of logical form output (noun/verb introduction, verb
relationships, and noun modifiers) are always computed but predicted
tokens for premature and completed phases are discarded and the next
predicted token set to that predicted by the current phase, decided by
the decoder based on comparing expected counts of nouns, verbs, and
relationships (computed from template matched by Encoder) to actual
counts in the output detected by decoder. Add clarifying comments.

Note: The original example and the new example are both simplified, but
the original example was even more simplified in that it was written to
ignore output order since ReCOGS semantic exact match is ok with that,
so for v_inf_taking to v_inf , it could be output e.g. "[nouns] [first
verb] [second verb] [agent for first verb] [xcomp v1 v2] [agent for
second verb]" (or other orderings of agent, xcomp, agent),
which means there are only 3 items after the noun/verb introduction
(which used to include outputting both verbs for v_inf), so a template
size of 3 (3 items after noun and verb introduction).

I am considering here to update the example to be consistent with
outputting it in the right order (at least in the basic case) as is done
by actual code, "[nouns] [first verb] [agent for first verb] [xcomp v1
v2] [second verb] [agent for second verb]", which means there are 4
items after the noun and verb introduction, or 5 if we treat v_inf as a
special case and do not introduce the first verb as part of noun/verb
introduction but count it separate.

For an example of RASP code for a similar problem that attempts to
output in the correct order always and is scored on that see the
separate https://github.com/willy-b/RASP-for-COGS project as the
original COGS task is scored by String Exact Match not Semantic Exact
Match. That implementation is also case-sensitive. But here we are
discussing a more loose model implemented at
https://github.com/willy-b/learning-rasp (note commit message is NOT intended to be deidentified, only the PDF from this branch should be safe in that way and will be submitted in 2025 May, NOT via GitHub).