Add additional v_dat_p2 generalization results notebook link to results notebooks already linked via colab in the arxiv version, to the deidentified version for arr review via their 4open.science mirror link (to be used in the 2025 May review; 4open.science mirror only used in that upcoming review, not previously), update to cite refereed version of papers instead of arXiv preprints for this version (#65)
(Note the preprint with my name on it is already up in arXiv, this is
not a secret paper, but I am preparing a version to facilitate blinded
review for a conference submission, where the submission window opens in
May. This is for that version.)

The results notebook link added here is for a 4open.science mirror of a
results notebook that has already been public and reachable from the
arXiv preprint version though that does not directly list it as I am
doing here (in the arXiv version, one must click the footnote link in
"9.9 v_dat_p2 recipient pp-modification for generalization assessment
and data augmentation attempt" and then navigate from there to the
v_dat_p2 results). I did not put it in the GitHub but left it in Colab
(intentionally).

Also, update this version for the Association for Computational
Linguistics to cite only refereed versions where possible instead of
original author preprints in arXiv per their policy at
https://web.archive.org/web/20250317191828/https://www.aclweb.org/adminwiki/index.php/ACL_Policies_for_Review_and_Citation .
.

PDF preview link: 

https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/51420e6c6cda751036c9d043ffb320cf2deb6ecb/rasp-for-recogs_pos-2024-draft.pdf