README.md updates to add a Results tables image and full Encoder-Decoder illustration (for #80 and #64) (#81)
New results snippet to be shown as part of README.md:
![RASP-for-ReCOGS-paper-main-results-page](https://github.com/user-attachments/assets/2dd19f30-8df7-45fb-a6e3-4f15e947ece4)

Note (Klinger et al 2024) (whose description of the COGS input grammar
was used to check that these flat patterns cover the non-recursive parts
of the grammar) are doing a CFG hierarchical parse of the input, we are
doing flat pattern matching after masking, with no hierarchical parsing.
We compare to them but are NOT using their approach.

README with changes preview link:
https://github.com/willy-b/RASP-for-ReCOGS/blob/91145238774ab0dc3152cf98970d819a0ce6dcad/README.md

RASP-for-ReCOGS Encoder-Decoder loop illustration (Figure 7 ref in image
above):
- figure image (in this PR):
https://github.com/willy-b/RASP-for-ReCOGS/blob/e639e5defb11acd85665dea2f0f249e5e2e4e8d1/rasp-for-recogs-decoder-loop-supplementary-figure-incl-encoder-and-decoder-and-grammar-patterns-vertical-layout.svg
- in context:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/8b59ec667df3ad8baccbb01963998b504b9043ba/rasp-for-recogs_pos-2024-draft.pdf#figure.caption.13

pp-masking and unrolling depth 2 pp recursion in the Decoder loop
(Figure 8 ref from image):
- figure image (in this PR):
https://github.com/willy-b/RASP-for-ReCOGS/blob/e639e5defb11acd85665dea2f0f249e5e2e4e8d1/rasp-for-recogs-decoder-loop-supplementary-figure_pp_depth_2.svg
- in context:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/8b59ec667df3ad8baccbb01963998b504b9043ba/rasp-for-recogs_pos-2024-draft.pdf#figure.caption.14

cp-masking and unrolling depth 2 cp recursion in the Decoder loop
(Figure 9 ref from image):
- figure image (in this PR):
https://github.com/willy-b/RASP-for-ReCOGS/blob/e639e5defb11acd85665dea2f0f249e5e2e4e8d1/rasp-for-recogs-decoder-loop-supplementary-figure_cp_depth_2.svg
- in context:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/8b59ec667df3ad8baccbb01963998b504b9043ba/rasp-for-recogs_pos-2024-draft.pdf#figure.caption.15