cite Friedman et al 2025 and do some small text cleanup
July 2024 version of Friedman et al in arXiv (https://arxiv.org/abs/2407.10949v1) was read by me and helped motivate that implementing semantic parsing as operationalized by ReCOGS by hand in RASP would be feasible. As of 2025-04 it has been published at https://aclanthology.org/2025.naacl-long.163/ , so cite it there (was not available when version submitted for review to Association for Computational Linguistics in 2025-05 was prepared).

Remove "(length and recursion depth limited)" in a description of ReCOGS in the introduction as it is confusing and how RASP-for-ReCOGS is consistent with length and recursion depth generalization is discussed in the Analysis section (we are not memorizing fixed depth shortcuts and using depth-general rules for unrolling recursion in the Decoder loop). That the ReCOGS recursion splits are fixed depth is obvious and also discussed elsewhere.