Cite Friedman et al 2025 and do some small text cleanup (#85)
PDF preview link:
https://raw.githubusercontent.com/willy-b/RASP-for-ReCOGS/35053fcbd28889feec751789d73b574e1683acb9/rasp-for-recogs_pos-2024-draft.pdf

The July 2024 version of Friedman et al's "Representing Rule-based
Chatbots with Transformers" in arXiv
(https://arxiv.org/abs/2407.10949v1) was read by me before starting the
RASP-for-ReCOGS project and it helped motivate that implementing
semantic parsing as operationalized by ReCOGS (and COGS) by hand in RASP
would be feasible. However, I did not use anything specifically from
that paper and it was pre-peer review at that time so it was not cited
in the initial version of this paper. As of 2025-04 Friedman et al 2025
has been published at https://aclanthology.org/2025.naacl-long.163/ ,
and I have freed up space in the paper to add a citation, so now I will
cite it there (the peer reviewed form was not yet available when the
version of this paper (RASP-for-ReCOGS) submitted for review to
Association for Computational Linguistics Rolling Review in 2025-05 was
prepared).

Also remove "(length and recursion depth limited)" in a description of
ReCOGS in the introduction as it was confusing a reviewer that we were
implementing a fundamentally length/depth limited approach, and since
how RASP-for-ReCOGS is consistent with length and recursion depth
generalization is discussed in the Analysis section (we are not
memorizing fixed depth shortcuts and using depth-general rules for
unrolling recursion in the Decoder loop). That the ReCOGS recursion
splits are fixed depth is obvious and also discussed elsewhere.